refactor: Rename module with tests of internal async client
And extract ot conftest things that will be common for both internal clients.
